Move context factory
