+++ test/routes/entry-point.test.ts
@@ -2,7 +2,7 @@ import jsonld from 'jsonld';
 import { Next, Response } from 'koa';
 import entryPoint from '../../src/routes/entry-point';
 import runMiddleware from '../middleware';
-import createContext from './context';
+import createContext from '../context';
 
 const makeRequest = async (next?: Next): Promise<Response> => {
   const context = createContext();
